fix: window overrlaps polybar
my polybar config, in the bar module :

override-redirect = true

This line tells polybar it shouldn't "push" windows down when it appears.

Set the top padding in your bspwmrc file to whatever you like and this should do the trick. I have it set to $BAR_HEIGHT + $GAP, so that my window won't overlap the bar